update product api and staff service

+++ frontend/web/src/services/product.service.ts
@@ -3,21 +3,29 @@ import { apiClient } from "./api";
 
 export const productService = {
   getAllProduct: async (params: ProductFilterParams) => {
-    const response = await apiClient.get(`/products`, {
+    const response = await apiClient.get(`/products/admin`, {
       params,
     });
     return response.data;
   },
   addProduct: async (data: FormData) => {
-    const response = await apiClient.post(`/products`, data);
+    const response = await apiClient.post(`/products/admin`, data, {
+      headers: {
+        "Content-Type": "multipart/form-data",
+      }
+    });
     return response.data;
   },
   deleteProduct: async (id: string) => {
-    const response = await apiClient.delete(`/products/${id}`);
+    const response = await apiClient.delete(`/products/admin/${id}`);
     return response.data;
   },
   updateProduct: async (id: string, data: FormData) => {
-    const response = await apiClient.put(`/products/${id}`, data);
+    const response = await apiClient.put(`/products/admin/${id}`, data, {
+      headers: {
+        "Content-Type": "multipart/form-data",
+      },
+    });
     return response.data;
   },
 };
